Move project service concerns to service;
Update overwrite codes utility with correct label value keyname indicator
